+++ pages/api/articles/reply/index.js
@@ -1,5 +1,6 @@
 import connectMongoDB from "@/lib/services/database/mongodb";
 import ReplyArticleComment from "@/database/models/ReplyArticleComment";
+import ArticleComments from "@/database/models/ArticleComments";
 
 export default async function handler(req, res) {
     const { method } = req;
@@ -9,8 +10,9 @@ export default async function handler(req, res) {
     switch (method) {
         case 'POST':
             try {
-                const { commentId, userId, reply } = req.body;
-                const newReply = await ReplyArticleComment.create({ commentId, userId, reply });
+                const newReply = await ReplyArticleComment.create(req.body);
+                await ArticleComments.findByIdAndUpdate(req.body.commentId, 
+                    { $push: { reply: newReply._id } });
                 res.status(201).json({ success: true, data: newReply });
             } catch (error) {
                 res.status(400).json({ success: false, error: error.message });
@@ -30,8 +32,18 @@ export default async function handler(req, res) {
         case 'DELETE':
             try {
                 const { replyId } = req.query;
+                const reply = await ReplyArticleComment.findById(replyId);
+                if (!reply) {
+                    return res.status(400).json({ success: false });
+                }
+                
+                await ArticleComments.findByIdAndUpdate(reply.commentId, 
+                    { $pull: { reply: replyId } 
+                });
+
                 await ReplyArticleComment.findByIdAndDelete(replyId);
-                res.status(200).json({ success: true });
+
+                res.status(200).json({ success: true, data: {} });
             } catch (error) {
                 res.status(400).json({ success: false, error: error.message });
             }
